refactor: DiaryContentLayout에서 의미없는 depth 줄이기
