+++ fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryComponent/DiaryContentsLayout/DiaryContentsLayout.styled.ts
@@ -4,5 +4,28 @@ import { StyleSheet } from "react-native";
 export const styles = StyleSheet.create({
     layout: {
         marginTop: 15
-    }
+    },
+    container: {
+        marginLeft: 20,
+        fontSize: 30,
+    },
+    content: {
+        fontSize: 15,
+        fontStyle: 'normal'
+    },
+    button: {
+        padding: 5,
+        borderRadius: 5,
+        borderWidth: 1,
+        borderColor: 'gray',
+        backgroundColor: 'gray',
+        textAlign: 'center',
+        marginLeft: 5,
+    },
+    more: {
+        backgroundColor: 'white',
+    },
+    moreText: {
+        color: 'gray',
+    },
 });
