+++ fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryComponent/DiaryContentsLayout/DiaryContentsLayout.tsx
@@ -1,8 +1,8 @@
-import React from "react";
-import DiaryContent from "./DiaryContent/DiaryContent";
-import { LocalDate, LocalDateTime } from "@js-joda/core";
-import { View } from "react-native";
+import React, { useState } from "react";
+import { Text, View } from "react-native";
 import * as S from "./DiaryContentsLayout.styled";
+import Button from "../../../../components/common/Button/Button";
+import { diaryContentPreviewLines, diaryContentPreviewWordCount } from "../../../../constants/diary";
 
 
 export interface DiaryContentsLayoutProps {
@@ -10,11 +10,31 @@ export interface DiaryContentsLayoutProps {
 }
 
 const DiaryContentsLayout = (props: DiaryContentsLayoutProps) => {
-    const { content } = props;
+    const { content, ...rest } = props;
+    const [isExpanded, setIsExpanded] = useState(false);
+
+    const shouldShowReadMore = content.length > diaryContentPreviewWordCount || content.split("\n").length > diaryContentPreviewLines;
+
+    const handleToggleExpand = () => {
+        setIsExpanded(!isExpanded);
+    };
 
     return (
         <View style={S.styles.layout}>
-            <DiaryContent content={content}/>
+            <View style={S.styles.container}>
+                <Text style={S.styles.content} numberOfLines={isExpanded ? diaryContentPreviewLines : content.split("\n").length}>
+                    {isExpanded ? content : content.slice(0, diaryContentPreviewWordCount)}
+                    {!isExpanded && content.length > diaryContentPreviewWordCount && '...'}
+                </Text>
+                {shouldShowReadMore && (
+                    <Button
+                        title={isExpanded ? "간단히 보기" : "더보기"}
+                        onPress={handleToggleExpand}
+                        style={S.styles.more}
+                        textStyle={S.styles.moreText}
+                    />
+                )}
+            </View>
         </View>
     );
 }
